+++ Controllers/FacultyController.cs
@@ -17,41 +17,72 @@ namespace PRMS.Controllers
         // GET: /Faculty/
         public ActionResult Index()
         {
+            if (!HasSession()) return RedirectToAction("Index", "Home");
+
             return RedirectToAction("AddFaculty", "Faculty");
+        }
+
+        protected Boolean HasSession()
+        {
+            Admin admin = HttpContext.Session[Variables.AdminSession] as Admin;
+            if (admin == null)
+            {
+                return false;
+            }
+
+            return true;
         }
 
 
         public ActionResult AddFaculty()
         {
+            if (!HasSession()) return RedirectToAction("Index", "Home");
+
             return View();
         }
+
         [HttpPost]
         public ActionResult AddFaculty([Bind(Include = "FacultyName,ShortForm")] Faculty faculty)
         {
-            CreateDatabaseAndTable cdt = new CreateDatabaseAndTable(faculty.ShortForm);
-            cdt.CreateDatabase(faculty.ShortForm);
+            if (!HasSession()) return RedirectToAction("Index", "Home");
 
-            db.Faculty.Add(faculty);
-            db.SaveChanges();
+            faculty.ShortForm = faculty.ShortForm.ToUpper();
+
+            if (db.Faculty.Where(a => a.ShortForm == faculty.ShortForm || a.FacultyName == faculty.FacultyName).FirstOrDefault() == null)
+            {
+                CreateDatabaseAndTable cdt = new CreateDatabaseAndTable(faculty.ShortForm);
+                cdt.CreateDatabase(faculty.ShortForm);
+
+                db.Faculty.Add(faculty);
+                db.SaveChanges();
+                ViewBag.Message = "Successfully Faculty Added..! ";
+            }
+            else
+            {
+                ViewBag.ErrorMessage = "Faculty Already Exists..!"; 
+            }
 
             return View();
         }
 
         public ActionResult ManageFaculty()
         {
+            if (!HasSession()) return RedirectToAction("Index", "Home");
+
             return View(db.Faculty.ToList());
 
         }
         public ActionResult EditFaculty(int? id)
         {
+            if (!HasSession()) return RedirectToAction("Index", "Home");
 
-            //  return View("~/Views/Admin/EditTeacher.cshtml");
             return RedirectToAction("EditFaculties", "Faculty", new { ind = id });
-
-           // return View(db.Faculty.Find(id));
         }
+
         public ActionResult EditFaculties(int? ind)
         {
+            if (!HasSession()) return RedirectToAction("Index", "Home");
+
             if (ind == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -69,7 +100,21 @@ namespace PRMS.Controllers
         [HttpPost]
         public ActionResult EditFaculties([Bind(Include = "id,FacultyName,ShortForm")] Faculty faculty)
         {
-            
+            if (!HasSession()) return RedirectToAction("Index", "Home");
+
+            faculty.ShortForm = faculty.ShortForm.ToUpper();
+
+            Faculty oldFaculty = db.Faculty.Find(faculty.id);
+            db.Entry(oldFaculty).State = EntityState.Detached;
+            Boolean ck=false;
+            if (!faculty.ShortForm.Equals(oldFaculty.ShortForm))
+            {
+                CreateDatabaseAndTable cdt = new CreateDatabaseAndTable(oldFaculty.ShortForm);
+                  ck=  cdt.AlterDatabase(oldFaculty.ShortForm, faculty.ShortForm);
+            }
+           
+            if (ck == false) return View(db.Faculty.Find(faculty.id));
+
             if (ModelState.IsValid)
             {
                 db.Entry(faculty).State = EntityState.Modified;
@@ -82,6 +127,8 @@ namespace PRMS.Controllers
 
         public ActionResult DeleteFaculty(int? id)
         {
+            if (!HasSession()) return RedirectToAction("Index", "Home");
+
             Faculty faculty = db.Faculty.Find(id);
             db.Faculty.Remove(faculty);
             db.SaveChanges();
